backend/requirements.txt
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
PyPDF2==3.0.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
